Refactored the Orders_Test and moved into PayPal_Orders_test and Stripe_Orders_test. Moved some logic into traint Fee_Creator.
